Move `nativeCurrencyTypes` from helpers to entities.
